emacs: disable switching of <return> and C-j for ivy
When there are two entries and one is a prefix of the other, this allows enter to open the selected
one instead of the one whose name is typed (e.g. typing '.zsh' shows a directory .zsh and a file
.zshrc, previously no matter which the user selected `counsel-find-file` would open the directory
since the name is an exact match).

The reason for switching is abo-abo/swiper#1102 causing <return> on a directory to open it in dired
instead of going down that directory.