Add more LSP bindings
lsp-rename is not a part of lsp-ui, but lsp itself. However, lsp itself does not define a key map
like lsp ui does.

There's no `ensure` at the use package of `lsp-ui-flycheck` because this module is already installed
as a part of `lsp-ui`. However, there seem to be an issue with auto loading `lsp-ui-flycheck-list`
so this is necessary.